add .gitignore add new AddDispatchR extensions that allow passing in a ConfigurationOptions object add support for scanning multiple assemblies for types
