+++ src/DispatchR/Configuration/ConfigurationOptions.cs
@@ -1,0 +1,11 @@
+﻿using System.Reflection;
+
+namespace DispatchR.Configuration
+{
+    public class ConfigurationOptions
+    {
+        public bool RegisterPipelines { get; set; } = true;
+        public bool RegisterNotifications { get; set; } = true;
+        public List<Assembly> Assemblies { get; } = new();
+    }
+}
